Added zeroshot/separate BERTopic_model from train
