+++ topic_modelling/bert/bert_topic_config.yml
@@ -7,6 +7,7 @@ model_0:
   ctfidf_params: {"bm25_weighting": True, "reduce_frequent_words": True}
   representation_model: KeyBERTInspired()
   min_topic_size: 30
+  nr_topics: 'auto'
 
 model_1:
   model_name: 'sentence-transformers-key-unigram'
@@ -17,6 +18,7 @@ model_1:
   ctfidf_params: {"bm25_weighting": True, "reduce_frequent_words": True}
   representation_model: KeyBERTInspired()
   min_topic_size: 30
+  nr_topics: 'auto'
 
 model_2:
   model_name: 'bert-base-uncased-key-unigram'
@@ -27,6 +29,7 @@ model_2:
   ctfidf_params: {"bm25_weighting": True, "reduce_frequent_words": True}
   representation_model: KeyBERTInspired()
   min_topic_size: 30
+  nr_topics: 'auto'
 
 model_3:
   model_name: 'bert-base-uncased-unigram'
@@ -37,6 +40,7 @@ model_3:
   ctfidf_params: {"bm25_weighting": True, "reduce_frequent_words": True}
   representation_model: None
   min_topic_size: 30
+  nr_topics: 'auto'
 
 model_4:
   model_name: 'sentence-transformers-key-bigram-v2'
@@ -47,6 +51,18 @@ model_4:
   ctfidf_params: {"bm25_weighting": True, "reduce_frequent_words": False}
   representation_model: KeyBERTInspired()
   min_topic_size: 30
+  nr_topics: 'auto'
+
+model_5:
+  model_name: 'zero-shot'
+  embedding_model: SentenceTransformer('all-MiniLM-L12-v2')
+  dim_reduction_model: None
+  clustering_model: None
+  vectorizer_params: {'min_df': 1, 'max_df': 1.0, "ngram_range": 2}
+  ctfidf_params: {"bm25_weighting": True, "reduce_frequent_words": False}
+  representation_model: ZeroShotClassification(['drinks', 'cat', 'dog', 'carbohydrates', 'chocolate', 'snacks', 'healthy alternatives', 'household', 'baby food', 'sauce'], model="facebook/bart-large-mnli", min_prob = 0.5)
+  min_topic_size: 30
+  nr_topics: None
 
 data_folder: data/reviews.csv
 image_folder: image/
